# server/swagger.yaml
openapi: 3.0.0
info:
  title: App_Vaa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /events/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Events
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Events
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
    post:
      summary: CRUD ACTION update
      tags:
        - Events
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Events"
  /events/findBy_clubs/{key}:
    get:
      summary: CRUD ACTION findBy_clubs
      tags:
        - Events
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Events"
    
  #-----  States  -----
...
  /students/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Students
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Students
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
    post:
      summary: CRUD ACTION update
      tags:
        - Students
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Students"
  /students/findBy_classes/{key}:
    get:
      summary: CRUD ACTION findBy_classes
      tags:
        - Students
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Students"
  /students/findBy_clubs/{key}:
    get:
      summary: CRUD ACTION findBy_clubs
      tags:
        - Students
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Students"
    
  #-----  Teachers  -----
...
  /teachers/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
    post:
      summary: CRUD ACTION update
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teachers"
  /teachers/findBy_classes/{key}:
    get:
      summary: CRUD ACTION findBy_classes
      tags:
        - Teachers
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
  /teachers/findBy_clubs/{key}:
    get:
      summary: CRUD ACTION findBy_clubs
      tags:
        - Teachers
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Events:
      properties:
        _id:
          type: string
        name:
          type: string
        vacancies:
          type: number
      required:
        -  _id
        - name
        - vacancies
...
    Students:
      properties:
        _id:
          type: string
        date_of_birth:
          type: integer
        facebook:
          type: string
        instagram:
          type: string
        name:
          type: string
        password:
          type: string
        position_secundary:
          type: number
        position_usual:
          type: number
        steer:
          type: boolean
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - date_of_birth
        - name
        - password
        - position_secundary
        - position_usual
        - steer
        - surname
        - username
    Teachers:
      properties:
        _id:
          type: string
        auth_classes_students:
          type: boolean
        date_of_birth:
          type: integer
        name:
          type: string
        partner:
          type: boolean
        password:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - auth_classes_students
        - date_of_birth
        - name
        - partner
        - password
        - surname
        - username
...